backend/pyproject.toml
[project]
name = "lh-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.1.0",
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "python-dotenv",
    "httpx",
]
